Added delete directive, changed DHT connection mechanism
- Due to `git push -d` not sending a commit/blob we can sign, `sign` command now includes `-d` flag to create a commit that has a deletion directive for a target reference.

- The DHT service is no longer a reactor. It is now started manually and before tendermint core.

- Updated tendermint core version to `v0.32.6.fork.3` which allows certain tx response code to cause a block to be re-applied.

+++ types/errs.go
@@ -84,7 +84,7 @@ var (
 // Transaction processing errors
 const (
 	ErrCodeFailedDecode                = uint32(1)
-	ErrCodeExecFailure                 = 2
+	ErrCodeExecFailure                 = uint32(2)
 	ErrCodeMaxTxTypeReached            = 3
 	ErrCodeTxTypeUnexpected            = 4
 	ErrCodeTxInvalidValue              = 5
